requirements.txt
absl-py==1.0.0
appnope==0.1.3
asttokens==2.0.5
backcall==0.2.0
black==22.6.0
cfgv==3.3.1
chex==0.1.3
click==8.1.3
cloudpickle==2.1.0
cycler==0.11.0
debugpy==1.6.2
decorator==5.1.1
diffrax==0.1.1
distlib==0.3.4
distrax==0.1.2
dm-tree==0.1.7
einops==0.4.1
entrypoints==0.4
equinox==0.4.0
executing==0.9.1
filelock==3.7.1
flatbuffers==2.0
flax==0.4.1
fonttools==4.33.2
gast==0.5.3
ghp-import==2.1.0
griffe==0.22.0
identify==2.5.1
importlib-metadata==4.12.0
ipykernel==6.15.1
ipython==8.4.0
isort==5.10.1
jaxopt==0.4.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jupyter-client==7.3.4
jupyter-core==4.11.1
kiwisolver==1.4.2
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mergedeep==1.3.4
mkdocs==1.3.1
mkdocs-autorefs==0.4.1
mkdocs-material==8.3.9
mkdocs-material-extensions==1.0.3
mkdocstrings==0.19.0
mkdocstrings-python==0.7.1
msgpack==1.0.3
nest-asyncio==1.5.5
nodeenv==1.7.0
numpy==1.22.3
opt-einsum==3.3.0
optax==0.1.2
packaging==21.3
pandas==1.4.2
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
platformdirs==2.5.2
pre-commit==2.20.0
prompt-toolkit==3.0.30
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.12.0
pymdown-extensions==9.5
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
pyupgrade==2.37.1
PyYAML==6.0
pyyaml_env_tag==0.1
pyzmq==23.2.0
scikit-learn==1.1.1
scipy==1.8.0
seaborn==0.11.2
six==1.16.0
sklearn==0.0
stack-data==0.3.0
tensorflow-probability==0.17.0
threadpoolctl==3.1.0
tokenize-rt==4.2.1
toml==0.10.2
tomli==2.0.1
toolz==0.11.2
tornado==6.2
traitlets==5.3.0
typing==3.7.4.3
typing_extensions==4.2.0
virtualenv==20.15.1
watchdog==2.1.9
wcwidth==0.2.5
zipp==3.8.1
